Address PR review: move helpers to parser.cc, fix feature-test macro, add errno checks
