+++ onnx/test/cpp/parser_test.cc
@@ -836,12 +836,6 @@ TEST(ParserTest, LocaleIndependentFloatParsing) {
   EXPECT_EQ(node.attribute(0).name(), "alpha");
   float alpha = node.attribute(0).f();
   EXPECT_NEAR(alpha, 0.123f, 1e-6f) << "Float attribute misparsed under non-US locale";
-
-  // Also test the locale-independent conversion helpers directly.
-  EXPECT_NEAR(LocaleIndependentStof("3.14159"), 3.14159f, 1e-4f)
-      << "LocaleIndependentStof misparsed under non-US locale";
-  EXPECT_NEAR(LocaleIndependentStod("2.71828"), 2.71828, 1e-10)
-      << "LocaleIndependentStod misparsed under non-US locale";
 }
 
 } // namespace Test
